Change `web.` to `packages.`, `Registry` to `Repository`…
…so that all the links to the package registry use the new URLs, and all references to the registry call it the same thing.

We use `registry` and `repository` interchangeably, but they're different. What we're hosting is a registry — works almost exactly like NPM's package registry, but has Pulsar-specific packages.

It could be argued that it's also a “repository” of packages — not in the `git` sense, but just in the sense of the common noun. But “registry” functions equally well whether or not you know it's an NPM term of art, and doesn't run the risk of getting confused with any `git` terms of art.